requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
fastapi-cache2 
redis           #cache

#pip3 install -r requirements.txt   # to install it all above mentioned

# to create virtual environment
#python -m venv blog-env